Fix Getopt::Long issue with use bigint on perl < 5.10
The issue was that Getopt::Long checks default option type, and was choking
on getting a Math::BigInt instead of 1, because on perl < 5.10 no bigint
doesn't seem to do anything.

The fix is to use Math::BigInt exactly where we need it, and don't rely
on perl automagic bigint system.
This proves to be simpler and much less error prone.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>