Fix error when innodb values are not accessible
When mysql runs without InnoDB, mysql-snmp was not getting the innodb values.
But when the snmp client was requesting them we were not returning a
value, and the snmp client was failing.

This patch fixes this problem by jumping over unavailable values, returning
errors only if it is requested an unknown oid.

Bug reported by Scott Sanders.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>